+++ app/src/main/java/com/khasang/fixmynumber/Activity/ContactsListActivity.java
@@ -1,24 +1,25 @@
-package com.khasang.fixmynumber;
+package com.khasang.fixmynumber.Activity;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
-import android.widget.ListView;
 import android.widget.TextView;
 
-import org.w3c.dom.Text;
+import com.khasang.fixmynumber.Adapter.ContactsListAdapter;
+import com.khasang.fixmynumber.Model.ContactItem;
+import com.khasang.fixmynumber.R;
 
 import java.util.ArrayList;
 
-public class MainActivity extends AppCompatActivity {
+public class ContactsListActivity extends AppCompatActivity {
     ArrayList<ContactItem> contactList;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_contacts_list);
 
         createDummyContacts();
         setUpRecyclerView();
