+++ app/src/main/res/layout/activity_contacts_list.xml
